make chargedisplay a int and cast it as such

+++ src/game/client/ui/ms/vgui_health.h
@@ -108,7 +108,7 @@ namespace PrimaryHUD
 
 			m_Image.SetFrame(frame);
 
-			m_Label->setText(UTIL_VarArgs("%i/%i ", (int)m_CurrentAmt, (int)MaxAmt)); //the space is intentional
+			m_Label->setText( UTIL_VarArgs("%i/%i ", static_cast<int>(m_CurrentAmt), static_cast<int>(MaxAmt)) ); //the space is intentional
 			m_Label->SetFGColorRGB(COLOR(255, 255, 255, 10));
 			if (m_Type != 2) {
 				if (m_CurrentAmt > MaxAmt / 4.0f)
@@ -126,7 +126,7 @@ namespace PrimaryHUD
 	private:
 		float vChargeLevelAmt = 0;
 		float vCurChargeAmt = 0;
-		float vDisplayChargeLevel = 0;
+		int vDisplayChargeLevel = 0;
 		int mCurChargeLevel = 1;
 		int vCurChargeLevel = 0;
 
@@ -279,8 +279,8 @@ namespace PrimaryHUD
 						vChargeLevel++;
 					}
 					
-					vCurChargeLevel = (int)((vChargeLevel - 1) + vCurChargeAmt);
-					vDisplayChargeLevel = vCurChargeLevel - 1;
+					vCurChargeLevel = static_cast<int>(((vChargeLevel - 1) + vCurChargeAmt));
+					vDisplayChargeLevel = static_cast<int>(vCurChargeLevel - 1);
 
 					if(vCurChargeLevel > mCurChargeLevel)
 						PlayHUDSound(gEngfuncs.pfnGetCvarString("ms_chargebar_sound"), gEngfuncs.pfnGetCvarFloat("ms_chargebar_volume"));
